Ints should be returned as int64

+++ sql/Conn.go
@@ -27,15 +27,15 @@ type Tx interface {
 	SelectOne(output interface{}, query string, args ...interface{}) (err error)
 	SelectMaybe(output interface{}, query string, args ...interface{}) (err error)
 
-	SelectInt(query string, args ...interface{}) (num int, err error)
+	SelectInt(query string, args ...interface{}) (num int64, err error)
 	SelectString(query string, args ...interface{}) (str string, err error)
 	SelectUint(query string, args ...interface{}) (num uint, err error)
 
-	SelectIntMaybe(query string, args ...interface{}) (num int, found bool, err error)
+	SelectIntMaybe(query string, args ...interface{}) (num int64, found bool, err error)
 	SelectStringMaybe(query string, args ...interface{}) (str string, found bool, err error)
 	SelectUintMaybe(query string, args ...interface{}) (num uint, found bool, err error)
 
-	SelectIntForce(query string, args ...interface{}) (num int, err error)
+	SelectIntForce(query string, args ...interface{}) (num int64, err error)
 	SelectStringForce(query string, args ...interface{}) (str string, err error)
 	SelectUintForce(query string, args ...interface{}) (num uint, err error)
 }
